[Post,Comment,Great] controller - request dto 수정 1. userId 를 @RequestParam 으로 수정 2. 관련 메소드 수정

+++ src/main/java/com/example/board1/service/CommentService.java
@@ -5,11 +5,9 @@ import com.example.board1.dto.CommentResponseDto;
 import com.example.board1.entity.Comment;
 import com.example.board1.entity.Post;
 import com.example.board1.entity.User;
-import com.example.board1.exception.InvalidRequestException;
 import com.example.board1.exception.NotFoundException;
 import com.example.board1.repository.CommentRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -37,17 +35,18 @@ public class CommentService {
 
 
     @Transactional
-    public Comment createComment(CommentRequestDto dto) {
-        if (dto == null) {
-            throw new IllegalArgumentException("Comment ID and DTO must not be null");
-        }
-        Comment comment = this.toEntity(dto);
+    public Comment createComment(CommentRequestDto dto, String userId) {
+        User user = userService.findUserByUserId(userId);
+        Post post = postService.findPostByPostId(dto.getPostId());
+
+        Comment comment = this.toEntity(dto, userId);
+
         return commentRepository.save(comment);
     }
 
-    public Comment toEntity(CommentRequestDto dto) {
+    public Comment toEntity(CommentRequestDto dto, String userId) {
         Post post = postService.findPostByPostId(dto.getPostId());
-        User user = userService.findUserByUserId(dto.getUserId());
+        User user = userService.findUserByUserId(userId);
 
         return Comment.builder()
                 .content(dto.getContent())
@@ -65,10 +64,10 @@ public class CommentService {
 
 
     @Transactional
-    public Comment updateCommentById(Long commentId, CommentRequestDto dto) {
+    public Comment updateCommentById(Long commentId, CommentRequestDto dto, String userId) {
         Comment comment = this.findByCommentId(commentId);
 
-        User user = userService.findUserByUserId(dto.getUserId());
+        User user = userService.findUserByUserId(userId);
         User commentUser = this.findByCommentId(commentId).getUser();
         if (!user.equals(commentUser)) {
             throw new IllegalArgumentException("User not found for given userId");
@@ -80,8 +79,13 @@ public class CommentService {
 
 
     @Transactional
-    public void deleteById(Long commentId) {
+    public void deleteById(Long commentId, String userId) {
+        User user = userService.findUserByUserId(userId);
         Comment comment = this.findByCommentId(commentId);
+
+        if(!user.equals(comment.getUser())) {
+            throw new IllegalArgumentException("User not found for given userId");
+        }
         commentRepository.delete(comment);
     }
